+++ code/compiler/builder.node/Builder/Builder.clean.mts
@@ -1,5 +1,5 @@
-import { fs, execa, t } from './common/index.mjs';
-import { Paths } from './Paths.mjs';
+import { fs, t } from '../common/index.mjs';
+import { Paths } from '../Paths.mjs';
 
 /**
  * Clean a module of transient build artifacts and temporary data.
