+++ code/compiler/builder.node/Builder/Builder.mts
@@ -1,9 +1,9 @@
 import { build } from './Builder.build.mjs';
 import { clean } from './Builder.clean.mjs';
 import { test } from './Builder.test.mjs';
-import { FindUtil as Find } from './common/index.mjs';
-import { Dependencies } from './operations/Dependencies.mjs';
-import { Paths } from './Paths.mjs';
+import { FindUtil as Find } from '../common/index.mjs';
+import { Dependencies } from '../op/Dependencies.mjs';
+import { Paths } from '../Paths.mjs';
 
 /**
  * ESM module builder.
