+++ code/compiler/builder.node/Builder/Builder.test.mts
@@ -1,4 +1,4 @@
-import { execa, t } from './common/index.mjs';
+import { execa, t } from '../common/index.mjs';
 import Vitest from 'vitest';
 
 /**
