+++ code/compiler/builder.node/index.mts
@@ -1,2 +1,2 @@
 export { fs } from './common/index.mjs';
-export { Builder } from './Builder.mjs';
+export { Builder } from './Builder/index.mjs';
